# Project_Document/WatchCareAI/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WatchCare AI API
  version: 0.1.0
  description: >-
    守望 AI 云端 REST 接口（iOS App ↔ 云端），对应设计方案 5.2 节。
    边缘节点事件上行通道走 MQTT，见 contracts/mqtt/topics.md；
    /v1/events/ingest 仅为垂直切片阶段的开发旁路。
servers:
  - url: http://localhost:8000
    description: local dev
...
paths:
...
  /v1/events:
    get:
      tags: [events]
      summary: 查询风险事件（游标分页）
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: status
          in: query
          schema:
            type: string
            enum: [CREATED, NOTIFIED, ACKNOWLEDGED, RESOLVED, FALSE_POSITIVE, ESCALATED]
      responses:
        '200':
          description: 事件页
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '../schemas/risk-event-envelope.schema.json'
                  nextCursor: { type: [string, 'null'] }
  /v1/events/{eventId}:
    get:
      tags: [events]
      summary: 事件详情（含状态流转历史与处置记录）
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 事件详情 }
        '404': { description: 事件不存在 }
  /v1/events/{eventId}/acknowledgements:
    post:
      tags: [events]
      summary: 确认事件或标记误报（驱动状态机，见设计方案 4.3 节）
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgementRequest'
      responses:
        '200': { description: 状态已更新 }
        '409': { description: 非法状态迁移 }
  /v1/events/ingest:
    post:
      tags: [events]
      summary: 开发旁路 - 垂直切片阶段供边缘模拟器直接 POST 事件信封
      deprecated: true
      description: 生产环境事件上行必须走 MQTT（mTLS + QoS1），本接口仅用于本地联调。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/risk-event-envelope.schema.json'
      responses:
        '200': { description: 已存在（幂等命中） }
        '201': { description: 已创建 }
...
components:
  schemas:
    AcknowledgementRequest:
      type: object
      additionalProperties: false
      required: [action, actorId]
      properties:
        action:
          type: string
          enum: [NO_DANGER, CONFIRM_RISK, HANDLING, FALSE_ALARM]
        actorId: { type: string, minLength: 1 }
        reason: { type: string, default: '' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []